wrapped all run_query() statements into  'if execute' to avoid execution during compilation/parsing phase

+++ macros/ili_utils.sql
@@ -5,17 +5,18 @@
 --  call through run-operations e.g.:
 --  dbt run-operation create_ili_sequence --args '{schema: dbt_quellkataster}'
 {% macro create_ili_sequence(schema_name) -%}
-  -- create sequence
-  {% set query %}
-    CREATE SEQUENCE IF NOT EXISTS {{schema_name}}.t_ili2db_seq
-    INCREMENT 1
-    START 1
-    MINVALUE 1
-    MAXVALUE 9223372036854775807
-    CACHE 1;
-  {% endset %}
-  {% set query_return = run_query(query) %}
-
+  {% if execute %}
+    -- create sequence
+    {% set query %}
+      CREATE SEQUENCE IF NOT EXISTS {{schema_name}}.t_ili2db_seq
+      INCREMENT 1
+      START 1
+      MINVALUE 1
+      MAXVALUE 9223372036854775807
+      CACHE 1;
+    {% endset %}
+    {% set query_return = run_query(query) %}
+  {% endif %}
 {%- endmacro %}
 
 -- Reset 't_ili2db_seq' in target schema
@@ -31,19 +32,20 @@
   {{ log("Creating role " ~ schema_name ~ "_read", info=True) }}
   {{ log("Creating role " ~ schema_name ~ "_write", info=True) }}
 
-  {% set sql_query %}
-    CREATE ROLE {{schema_name}}_read;
-    CREATE ROLE {{schema_name}}_write;
-  {% endset %}
-  {% do run_query(sql_query) %}
+  {% if execute %}
+    {% set sql_query %}
+      CREATE ROLE {{schema_name}}_read;
+      CREATE ROLE {{schema_name}}_write;
+    {% endset %}
+    {% do run_query(sql_query) %}
 
-  {{ log("Assigning created roles to " ~ target.user, info=True) }}
-  {% set sql_query %}
-    GRANT {{schema_name}}_read to {{target.user}};
-    GRANT {{schema_name}}_write to {{target.user}};
-  {% endset %}
-  {% do run_query(sql_query) %}
-
+    {{ log("Assigning created roles to " ~ target.user, info=True) }}
+    {% set sql_query %}
+      GRANT {{schema_name}}_read to {{target.user}};
+      GRANT {{schema_name}}_write to {{target.user}};
+    {% endset %}
+    {% do run_query(sql_query) %}
+  {% endif %}
 {%- endmacro %}
 
 {% macro grant_select_on_all_tables(schema_name) -%}
@@ -58,11 +60,13 @@
     info=True
   )}}
 
-  {% set sql_query %}
-    GRANT USAGE ON SCHEMA {{schema_name}} TO {{target.user}};
-    GRANT ALL ON ALL TABLES IN SCHEMA {{schema_name}} TO {{target.user}};
-  {% endset %}
-  {% do run_query(sql_query) %}
+  {% if execute %}
+    {% set sql_query %}
+      GRANT USAGE ON SCHEMA {{schema_name}} TO {{target.user}};
+      GRANT ALL ON ALL TABLES IN SCHEMA {{schema_name}} TO {{target.user}};
+    {% endset %}
+    {% do run_query(sql_query) %}
+  {% endif %}
 {%- endmacro %}
 
 
@@ -70,75 +74,81 @@
 -- Populate basket table based on project configuration:
 -- Add a row (i.e.) a basket for each basket defined in dbt_project.yml
 {% macro setup_baskets(schema_name) %}
-  {% for key, value_dict in var('baskets').items() %}
-    {{ log(
-        "Writing " ~ key ~ " into " ~ schema_name ~ ".t_ili2db_basket",
-        info=True
-    )}}
-    {% set sql_basket_row %}
-      INSERT INTO {{schema_name}}.t_ili2db_basket(
-        t_id,         -- NOT NULL
-        dataset,
-        topic,        -- NOT NULL
-        t_ili_tid,
-        attachmentkey,-- NOT NULL (but also not used -> write '-')
-        domains
-      )
-      VALUES (
-        {{ value_dict['t_id'] }},
-        {{ value_dict['dataset_t_id'] }},
-        '{{ value_dict['topic'] }}',
-        uuid_generate_v4(),
-        '-',
-        NULL
-      );
-      -- advance t_ili2db_seq to make up for manually set t_id
-      SELECT nextval('{{schema_name}}.t_ili2db_seq'::regclass);
-    {% endset %}
-    {% set query_return = run_query(sql_basket_row)%}
-  {% endfor %}
+  {% if execute %}
+    {% for key, value_dict in var('baskets').items() %}
+      {{ log(
+          "Writing " ~ key ~ " into " ~ schema_name ~ ".t_ili2db_basket",
+          info=True
+      )}}
+      {% set sql_basket_row %}
+        INSERT INTO {{schema_name}}.t_ili2db_basket(
+          t_id,         -- NOT NULL
+          dataset,
+          topic,        -- NOT NULL
+          t_ili_tid,
+          attachmentkey,-- NOT NULL (but also not used -> write '-')
+          domains
+        )
+        VALUES (
+          {{ value_dict['t_id'] }},
+          {{ value_dict['dataset_t_id'] }},
+          '{{ value_dict['topic'] }}',
+          uuid_generate_v4(),
+          '-',
+          NULL
+        );
+        -- advance t_ili2db_seq to make up for manually set t_id
+        SELECT nextval('{{schema_name}}.t_ili2db_seq'::regclass);
+      {% endset %}
+      {% set query_return = run_query(sql_basket_row)%}
+    {% endfor %}
+  {% endif %}
 {%- endmacro %}
 
 {% macro setup_datasets(schema_name) %}
   -- Populate dataset table based on project configuration:
   -- Add a row (i.e.) a dataset for each dataset defined in dbt_project.yml
-  {% for key, value_dict in var('datasets').items() %}
-    {{ log(
-        "Writing " ~ key ~ " into " ~ schema_name ~ ".t_ili2db_dataset",
-        info=True
-    )}}
-    {% set sql_dataset_row %}
-      INSERT INTO {{schema_name}}.t_ili2db_dataset(
-        t_id,
-        datasetname
-      )
-      VALUES (
-        {{ value_dict['t_id'] }},
-        '{{ value_dict['datasetname'] }}'
-      );
-      -- advance t_ili2db_seq to make up for manually set t_id
-      SELECT nextval('{{schema_name}}.t_ili2db_seq'::regclass);
-    {% endset %}
-    {% set query_return = run_query(sql_dataset_row)%}
-  {% endfor %}
+  {% if execute %}
+    {% for key, value_dict in var('datasets').items() %}
+      {{ log(
+          "Writing " ~ key ~ " into " ~ schema_name ~ ".t_ili2db_dataset",
+          info=True
+      )}}
+      {% set sql_dataset_row %}
+        INSERT INTO {{schema_name}}.t_ili2db_dataset(
+          t_id,
+          datasetname
+        )
+        VALUES (
+          {{ value_dict['t_id'] }},
+          '{{ value_dict['datasetname'] }}'
+        );
+        -- advance t_ili2db_seq to make up for manually set t_id
+        SELECT nextval('{{schema_name}}.t_ili2db_seq'::regclass);
+      {% endset %}
+      {% set query_return = run_query(sql_dataset_row)%}
+    {% endfor %}
+  {% endif %}
 {%- endmacro %}
 
 {% macro reset_target_schema(schema_name) %}
-  {{ ili_utils.reset_ili_sequence(schema_name) }}
+  {% if execute %}
+    {{ ili_utils.reset_ili_sequence(schema_name) }}
 
-  -- Clear truncate tables
-  {% set sql_truncate %}
-    TRUNCATE TABLE {{schema_name}}.t_ili2db_dataset CASCADE;
-    -- Truncates:
-    -- t_ili2db_dataset, t_ili2db_basket, any data objects
-  {% endset %}
-  {% do run_query(sql_truncate) %}
+    -- Clear truncate tables
+    {% set sql_truncate %}
+      TRUNCATE TABLE {{schema_name}}.t_ili2db_dataset CASCADE;
+      -- Truncates:
+      -- t_ili2db_dataset, t_ili2db_basket, any data objects
+    {% endset %}
+    {% do run_query(sql_truncate) %}
 
-  -- Populate dataset table
-  {{ ili_utils.setup_datasets(schema_name) }}
+    -- Populate dataset table
+    {{ ili_utils.setup_datasets(schema_name) }}
 
-  -- Populate basket table
-  {{ ili_utils.setup_baskets(schema_name) }}
+    -- Populate basket table
+    {{ ili_utils.setup_baskets(schema_name) }}
+  {% endif %}
 {%- endmacro %}
 
 
@@ -151,28 +161,30 @@
       " with truncate_target=" ~ truncate_target, 
       info=true
     )}}
-  {% if truncate_target %}
-    {{ log(
-      "Truncating " ~ schema_name ~ "." ~ table_name,
-      info=true
-    )}}
-    {% set truncate_query %}
-      TRUNCATE TABLE {{schema_name}}.{{table_name}};
-    {% endset%}
-    {% set query_return = run_query(truncate_query)%}
-  {% endif %}
+  {% if execute %}
+    {% if truncate_target %}
+      {{ log(
+        "Truncating " ~ schema_name ~ "." ~ table_name,
+        info=true
+      )}}
+      {% set truncate_query %}
+        TRUNCATE TABLE {{schema_name}}.{{table_name}};
+      {% endset%}
+      {% set query_return = run_query(truncate_query)%}
+    {% endif %}
 
-  {% set insert_query %}
-    INSERT INTO {{schema_name}}.{{table_name}}(
-      -- Get list of column names present in boundary model
-      -- assumption: boundary model column names match target column names 
-      {{ dbt_utils.get_filtered_columns_in_relation(this) | join(',\n  ') }}
-    )
-    SELECT
-      *
-    FROM {{this}}
-  {% endset %}
-  {% set query_return = run_query(insert_query)%}
+    {% set insert_query %}
+      INSERT INTO {{schema_name}}.{{table_name}}(
+        -- Get list of column names present in boundary model
+        -- assumption: boundary model column names match target column names 
+        {{ dbt_utils.get_filtered_columns_in_relation(this) | join(',\n  ') }}
+      )
+      SELECT
+        *
+      FROM {{this}}
+    {% endset %}
+    {% set query_return = run_query(insert_query)%}
+  {% endif %}
 
 {%- endmacro %}
 
@@ -188,7 +200,7 @@
         info=True
     )}}
 
-    -- (try to) reset dbt schema's t_ili2db_seq
+    -- reset dbt schema's t_ili2db_seq
     {{ ili_utils.reset_ili_sequence(target.schema) }}
 
     {% if var('enable_transfer', false) %}
